+++ src/test/java/kr/hhplus/be/server/domain/point/UserPointFixture.kt
@@ -1,8 +1,10 @@
 package kr.hhplus.be.server.domain.point
 
+import java.math.BigDecimal
+
 class UserPointFixture(
     userId: Long = 1L,
-    point: Long = 100,
+    point: BigDecimal = BigDecimal.valueOf(100),
 ) : UserPoint(
     userId = userId,
     point = point,
